backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
spacy==3.7.2
rapidfuzz==3.6.1
openai==1.3.7
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
mypy==1.7.1
ruff==0.1.6 